pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-vllm"
version = "0.2.0"
authors = [{ name = "Xingkai Yu" }]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
description = "a lightweight vLLM implementation built from scratch"
requires-python = ">=3.10,<3.13"
dependencies = [
    "flashinfer-python>=0.6.8",
    "torch>=2.4.0",
    "triton>=3.0.0",
    "transformers>=4.51.0",
    "xxhash",
]

[project.optional-dependencies]
flash-attn = ["flash-attn"]
flashinfer = ["flashinfer-python>=0.6.8"]

[project.urls]
Homepage="https://github.com/GeeeekExplorer/nano-vllm"

[tool.setuptools.packages.find]
where = ["."]
include = ["nanovllm*"]
